+++ lib/index.d.ts
@@ -93,7 +93,8 @@ declare class SetMeUp {
     loadFromEnv(prefix?: string, options?: LoadEnvOptions): any;
     /**
      * Reset to default settings by unwatching and clearing all settings.
-     * Ideally you should call [[load]] / [[loadFromEnv]] after resetting.
+     * Ideally you should call [[load]] / [[loadFromEnv]] after resetting,
+     * otherwise it will most certainly break your application.
      * @event reset
      */
     reset(): void;